added server examples and started writing server code
