+++ .gitignore
@@ -128,3 +128,8 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+# model,data files
+*.npy
+*.h5
+*.csv
