+++ server/dev/machine_learning/input.txt
@@ -1,0 +1,1 @@
+무슨 일을 하든 흥미가 없고, 마음 한구석에 어둠이 번져있는 듯한 기분이었다. 끊임없이 떠오르는 부정적인 생각들을 물리치 며 하루를 마무리하려고 노력했지만, 마음 한편에는 여전히 어둠이 머물고 있었다.
