+++ server/dev/machine_learning/output.txt
@@ -1,0 +1,1 @@
+24.77
