--- server/dev/output.txt
@@ -1,1 +1,0 @@
-24.77
