+++ server/test_client.py
@@ -1,0 +1,12 @@
+import requests
+
+base_url = 'http://10.50.9.106:9999'
+
+url = base_url+'/Diary'
+body = {
+    'text' : "개미는 뚠뚠. 오늘도 뚠뚠. 열심히 일을 하네."
+}
+res = requests.post(url, json=body)
+
+data = res.json()
+print(data['score'])
